# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IDE Platform API — Replit-like workspace platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List all files in workspace as a tree
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileNode"

    post:
      operationId: createFile
      tags: [files]
      summary: Create a new file or folder
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileNode"
  /workspaces/{workspaceId}/files/read:
    post:
      operationId: readFileContent
      tags: [files]
      summary: Get file content by path (path in body)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilePathInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
  /workspaces/{workspaceId}/files/write:
    post:
      operationId: writeFileContent
      tags: [files]
      summary: Write/overwrite file content
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileContentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{workspaceId}/files/rename:
    post:
      operationId: renameFile
      tags: [files]
      summary: Rename or move a file/folder
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileRenameInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{workspaceId}/files/delete:
    post:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file or folder
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileDeleteInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── PORTS ──────────────────────────────────────────────────────────────────
  /workspaces/{workspaceId}/ports:
    get:
      operationId: listPorts
      tags: [ports]
      summary: List exposed ports
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Port"

    post:
      operationId: createPort
      tags: [ports]
      summary: Expose a port
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Port"
  /workspaces/{workspaceId}/ports/{portId}:
    delete:
      operationId: deletePort
      tags: [ports]
      summary: Remove an exposed port
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: portId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── SECRETS ────────────────────────────────────────────────────────────────
  /workspaces/{workspaceId}/secrets:
    get:
      operationId: listSecrets
      tags: [secrets]
      summary: List secrets (keys only, values masked)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Secret"

    post:
      operationId: createSecret
      tags: [secrets]
      summary: Create a secret
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
  /workspaces/{workspaceId}/secrets/{secretId}:
    patch:
      operationId: updateSecret
      tags: [secrets]
      summary: Update a secret value
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: secretId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"

    delete:
      operationId: deleteSecret
      tags: [secrets]
      summary: Delete a secret
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: secretId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── WORKFLOWS ──────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/domains:
    get:
      operationId: listDomains
      tags: [domains]
      summary: List domains attached to workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"

    post:
      operationId: createDomain
      tags: [domains]
      summary: Add a domain to workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
  /workspaces/{workspaceId}/domains/{domainId}:
    delete:
      operationId: deleteDomain
      tags: [domains]
      summary: Remove a domain
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: domainId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{workspaceId}/domains/{domainId}/verify:
    post:
      operationId: verifyDomain
      tags: [domains]
      summary: Verify DNS for a domain
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: domainId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"

  # ─── MONITORING ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ─── AUTH ────────────────────────────────────────────────────────────────
...
    FileNode:
      type: object
      required: [name, path, type]
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        size:
          type: ["number", "null"]
        children:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/FileNode"
    FileContent:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
        content:
          type: string
        language:
          type: ["string", "null"]
    FilePathInput:
      type: object
      required: [path]
      properties:
        path:
          type: string
    FileInput:
      type: object
      required: [path, type]
      properties:
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        content:
          type: ["string", "null"]
    FileContentUpdate:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
        content:
          type: string
    FileRenameInput:
      type: object
      required: [oldPath, newPath]
      properties:
        oldPath:
          type: string
        newPath:
          type: string
    FileDeleteInput:
      type: object
      required: [path]
      properties:
        path:
          type: string

    # ─── PORT ────────────────────────────────────────────────────────────────
    Port:
      type: object
      required: [id, workspaceId, internalPort, name, isPrivate, externalUrl]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        internalPort:
          type: number
        name:
          type: string
        isPrivate:
          type: boolean
        externalUrl:
          type: ["string", "null"]
        pid:
          type: ["number", "null"]
    PortInput:
      type: object
      required: [internalPort, name]
      properties:
        internalPort:
          type: number
        name:
          type: string
        isPrivate:
          type: boolean

    # ─── SECRET ──────────────────────────────────────────────────────────────
    Secret:
      type: object
      required: [id, workspaceId, key, createdAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        key:
          type: string
        createdAt:
          type: string
          format: date-time
    SecretInput:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
          minLength: 1
        value:
          type: string
    SecretUpdate:
      type: object
      required: [value]
      properties:
        value:
          type: string

    # ─── WORKFLOW ────────────────────────────────────────────────────────────
...
    Domain:
      type: object
      required: [id, workspaceId, domain, status, createdAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        domain:
          type: string
        status:
          type: string
          enum: [pending, active, error]
        sslStatus:
          type: ["string", "null"]
        dnsRecord:
          $ref: "#/components/schemas/DnsRecord"
        createdAt:
          type: string
          format: date-time
    DnsRecord:
      type: object
      required: [type, name, value]
      properties:
        type:
          type: string
        name:
          type: string
        value:
          type: string
    DomainInput:
      type: object
      required: [domain]
      properties:
        domain:
          type: string

    # ─── MONITORING ──────────────────────────────────────────────────────────
...