# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * IDE Platform API — Replit-like workspace platform
 * OpenAPI spec version: 0.1.0
 */
import type { DnsRecord } from './dnsRecord';
import type { DomainStatus } from './domainStatus';

export interface Domain {
  id: string;
  workspaceId: string;
  domain: string;
  status: DomainStatus;
  /** @nullable */
  sslStatus?: string | null;
  dnsRecord?: DnsRecord;
  createdAt: Date;
}